Move `tests` to the app's root
