+++ invariant.tests.ts
@@ -10,11 +10,11 @@ import {
   initializeClarityContext,
   initializeLocalContext,
   scheduleRendezvous,
-} from "../invariant";
+} from "./invariant";
 import {
   getFunctionsFromContractInterfaces,
   getSimnetDeployerContractsInterfaces,
-} from "../shared";
+} from "./shared";
 import { resolve } from "path";
 import fs from "fs";
 import fc from "fast-check";
@@ -23,8 +23,8 @@ import { Cl } from "@stacks/transactions";
 describe("File stream operations", () => {
   it("retrieves the invariant contract source", async () => {
     // Arrange
-    const manifestPath = resolve(__dirname, "../example/Clarinet.toml");
-    const contractsPath = resolve(__dirname, "../example/contracts");
+    const manifestPath = resolve(__dirname, "./example/Clarinet.toml");
+    const contractsPath = resolve(__dirname, "./example/contracts");
     const simnet = await initSimnet(manifestPath);
     const sutContractsInterfaces = getSimnetDeployerContractsInterfaces(simnet);
     const sutContractsList = Array.from(sutContractsInterfaces.keys());
@@ -51,8 +51,8 @@ describe("File stream operations", () => {
 describe("Simnet contracts operations", () => {
   it("retrieves Rendezvous contracts data", async () => {
     // Arrange
-    const manifestPath = resolve(__dirname, "../example/Clarinet.toml");
-    const contractsPath = resolve(__dirname, "../example/contracts");
+    const manifestPath = resolve(__dirname, "./example/Clarinet.toml");
+    const contractsPath = resolve(__dirname, "./example/contracts");
     const simnet = await initSimnet(manifestPath);
     const sutContractsInterfaces = getSimnetDeployerContractsInterfaces(simnet);
     const sutContractsList = Array.from(sutContractsInterfaces.keys());
@@ -87,8 +87,8 @@ describe("Simnet contracts operations", () => {
 
   it("deploys Rendezvous contracts to the simnet", async () => {
     // Arrange
-    const manifestPath = resolve(__dirname, "../example/Clarinet.toml");
-    const contractsPath = resolve(__dirname, "../example/contracts");
+    const manifestPath = resolve(__dirname, "./example/Clarinet.toml");
+    const contractsPath = resolve(__dirname, "./example/contracts");
     const simnet = await initSimnet(manifestPath);
     const sutContractsInterfaces = getSimnetDeployerContractsInterfaces(simnet);
     const sutContractsList = Array.from(sutContractsInterfaces.keys());
@@ -128,8 +128,8 @@ describe("Simnet contracts operations", () => {
 
   it("correctly filters the Rendezvous contracts interfaces", async () => {
     // Arrange
-    const manifestPath = resolve(__dirname, "../example/Clarinet.toml");
-    const contractsPath = resolve(__dirname, "../example/contracts");
+    const manifestPath = resolve(__dirname, "./example/Clarinet.toml");
+    const contractsPath = resolve(__dirname, "./example/contracts");
     const simnet = await initSimnet(manifestPath);
     const sutContractsInterfaces = getSimnetDeployerContractsInterfaces(simnet);
     const sutContractsList = Array.from(sutContractsInterfaces.keys());
@@ -160,7 +160,7 @@ describe("Simnet contracts operations", () => {
 
   it("correctly initializes the local context for a given functions map", async () => {
     // Arrange
-    const manifestPath = resolve(__dirname, "../example/Clarinet.toml");
+    const manifestPath = resolve(__dirname, "./example/Clarinet.toml");
     const simnet = await initSimnet(manifestPath);
     const sutContractsInterfaces = getSimnetDeployerContractsInterfaces(simnet);
     const sutContractsAllFunctions = getFunctionsFromContractInterfaces(
@@ -187,8 +187,8 @@ describe("Simnet contracts operations", () => {
 
   it("correctly initializes the Clarity context", async () => {
     // Arrange
-    const manifestPath = resolve(__dirname, "../example/Clarinet.toml");
-    const contractsPath = resolve(__dirname, "../example/contracts");
+    const manifestPath = resolve(__dirname, "./example/Clarinet.toml");
+    const contractsPath = resolve(__dirname, "./example/contracts");
     const simnet = await initSimnet(manifestPath);
     const sutContractsInterfaces = getSimnetDeployerContractsInterfaces(simnet);
     const sutContractsList = Array.from(sutContractsInterfaces.keys());
@@ -246,7 +246,7 @@ describe("Simnet contracts operations", () => {
 
   it("retrieves the contract source from the simnet", async () => {
     // Arrange
-    const manifestPath = resolve(__dirname, "../example/Clarinet.toml");
+    const manifestPath = resolve(__dirname, "./example/Clarinet.toml");
     const simnet = await initSimnet(manifestPath);
     const sutContractsInterfaces = getSimnetDeployerContractsInterfaces(simnet);
     const sutContractsList = Array.from(sutContractsInterfaces.keys());
